Split RecordContext file, add WithRecord documentation
