+++ docs/navigation.html
@@ -18,9 +18,12 @@
   <li {% if page.path contains 'ShowBase.md' %} class="active" {% endif %}><a class="nav-link" href="./ShowBase.html"><code>&lt;ShowBase&gt;</code></a></li>
   <li {% if page.path contains 'useShowController.md' %} class="active" {% endif %}><a class="nav-link" href="./useShowController.html"><code>useShowController</code></a></li>
   <li {% if page.path contains 'useShowContext.md' %} class="active" {% endif %}><a class="nav-link" href="./useShowContext.html"><code>useShowContext</code></a></li>
-  <li {% if page.path contains 'useRecordContext.md' %} class="active" {% endif %}><a class="nav-link" href="./useRecordContext.html"><code>useRecordContext</code></a></li>
 </ul>
 
+<ul><div>Common</div>
+  <li {% if page.path contains 'WithRecord.md' %} class="active" {% endif %}><a class="nav-link" href="./WithRecord.html"><code>&lt;WithRecord&gt;</code></a></li>
+  <li {% if page.path contains 'useRecordContext.md' %} class="active" {% endif %}><a class="nav-link" href="./useRecordContext.html"><code>useRecordContext</code></a></li>
+</ul>
 
 <ul><div>View Configuration</div>
   <li {% if page.path contains 'List.md' %} class="active" {% endif %}><a class="nav-link" href="./List.html"><code>&lt;List&gt;</code> View</a></li>
