+++ packages/ra-core/src/controller/RecordContext.tsx
@@ -1,112 +1,51 @@
 import * as React from 'react';
-import { createContext, ReactNode, ReactElement, useContext } from 'react';
+import { createContext, ReactNode } from 'react';
 import { Record } from '../types';
 
 /**
- * Context to store the current record.
+ * Context to store a record.
  *
- * Use the useRecordContext() hook to read the context. That's what the Edit and Show components do in react-admin.
- *
- * @example
- *
- * import { useEditController, EditContext } from 'ra-core';
- *
- * const Edit = props => {
- *     const { record } = useEditController(props);
- *     return (
- *         <RecordContextProvider value={record}>
- *             ...
- *         </RecordContextProvider>
- *     );
- * };
+ * @see RecordContextProvider
+ * @see useRecordContext
  */
 export const RecordContext = createContext<Record | Omit<Record, 'id'>>(
     undefined
 );
 
+RecordContext.displayName = 'RecordContext';
+
+/**
+ * Provider for the Record Context, to store a record.
+ *
+ * Use the useRecordContext() hook to read the context.
+ * That's what the Edit and Show components do in react-admin.
+ *
+ * Many react-admin components read the RecordContext, including all Field
+ * components.
+ *
+ * @example
+ *
+ * import { useGetOne, RecordContextProvider } from 'ra-core';
+ *
+ * const Show = ({ resource, id }) => {
+ *     const { data } = useGetOne(resource, id);
+ *     return (
+ *         <RecordContextProvider value={data}>
+ *             ...
+ *         </RecordContextProvider>
+ *     );
+ * };
+ */
 export const RecordContextProvider = <
     RecordType extends Record | Omit<Record, 'id'> = Record
 >({
     children,
     value,
-}: RecordContextOptions<RecordType>) => (
+}: RecordContextProviderProps<RecordType>) => (
     <RecordContext.Provider value={value}>{children}</RecordContext.Provider>
 );
 
-RecordContext.displayName = 'RecordContext';
-
-export interface RecordContextOptions<RecordType> {
+export interface RecordContextProviderProps<RecordType> {
     children: ReactNode;
     value?: RecordType;
-}
-
-/**
- * Hook to read the record from a RecordContext.
- *
- * Must be used within a <RecordContext> such as provided by the <EditContextProvider>
- * (e.g. as a descendent of <Edit> or <EditBase>) or within a <ShowContextProvider>
- * (e.g. as a descendent of <Show> or <ShowBase>)
- *
- * @example // basic usage
- *
- * import { useRecordContext } from 'ra-core';
- *
- * const TitleField = () => {
- *     const record = useRecordContext();
- *     return <span>{record && record.title}</span>;
- * };
- *
- * @example // allow record override via props
- *
- * import { useRecordContext } from 'ra-core';
- *
- * const TitleField = (props) => {
- *     const record = useRecordContext(props);
- *     return <span>{record && record.title}</span>;
- * };
- * render(<TextField record={record} />);
- *
- * @returns {Record} A record object
- */
-export const useRecordContext = <
-    RecordType extends Record | Omit<Record, 'id'> = Record
->(
-    props?: UseRecordContextParams<RecordType>
-): RecordType => {
-    // Can't find a way to specify the RecordType when CreateContext is declared
-    // @ts-ignore
-    const context = useContext<RecordType>(RecordContext);
-
-    return (props && props.record) || context;
-};
-
-export interface UseRecordContextParams<
-    RecordType extends Record | Omit<Record, 'id'> = Record
-> {
-    record?: RecordType;
-    [key: string]: any;
-}
-
-/**
- * Render prop version of useRecordContext
- *
- * @example
- * const BookShow = () => (
- *    <Show>
- *       <SimpleShowLayout>
- *          <WithRecord render={record => <span>{record.title}</span>} />
- *      </SimpleShowLayout>
- *   </Show>
- * );
- */
-export const WithRecord = <RecordType extends Record>({
-    render,
-}: WithRecordProps<RecordType>) => {
-    const record = useRecordContext<RecordType>();
-    return record ? render(record) : null;
-};
-
-export interface WithRecordProps<RecordType extends Record> {
-    render: (record: RecordType) => ReactElement;
-    label?: string;
 }
