+++ packages/ra-core/src/controller/index.ts
@@ -59,10 +59,12 @@ export {
     useListParams,
 };
 
-export * from './field';
-export * from './input';
 export * from './button';
 export * from './details';
+export * from './field';
+export * from './input';
 export * from './RecordContext';
 export * from './saveModifiers';
 export * from './useList';
+export * from './useRecordContext';
+export * from './WithRecord';
